move proptypes for uiSlides on organims review-slide
