+++ packages/@coorpacademy-components/src/template/app-review/player/prop-types.js
@@ -5,22 +5,6 @@ import ReviewCongratsPropTypes from '../../../organism/review-congrats/prop-type
 import ReviewBackgroundPropTypes from '../../../atom/review-background/prop-types';
 import ReviewHeader from '../../../organism/review-header/prop-types';
 import SlidePropTypes from '../../../organism/review-slide/prop-types';
-import AnswerPropTypes from '../../../molecule/answer/prop-types';
-
-const NextSlideProp = PropTypes.shape({
-  hidden: PropTypes.bool,
-  position: PropTypes.number,
-  animationType: PropTypes.string, // 'unstack' | 'restack'
-  isCorrect: PropTypes.boolean,
-  endReview: PropTypes.bool,
-  questionText: PropTypes.string,
-  answerUI: AnswerPropTypes
-});
-
-const SlideProp = PropTypes.shape({
-  ...NextSlideProp.props,
-  nextContent: NextSlideProp
-});
 
 const StepItemProp = PropTypes.shape({
   current: PropTypes.bool,
@@ -32,13 +16,7 @@ export const SlidesReviewPropTypes = {
   headerProps: PropTypes.shape(omit('steps', ReviewHeader.propTypes)),
   reviewBackgroundAriaLabel: ReviewBackgroundPropTypes['aria-label'],
   validate: SlidePropTypes.validate,
-  uiSlides: PropTypes.shape({
-    '0': SlideProp,
-    '1': SlideProp,
-    '2': SlideProp,
-    '3': SlideProp,
-    '4': SlideProp
-  }),
+  uiSlides: SlidePropTypes.uiSlides,
   apiSlides: SlidePropTypes.slides,
   finishedSlides: SlidePropTypes.finishedSlides,
   stepItems: PropTypes.shape({
